Fix the four bugs the http suite had pinned as characterization tests
Each of these was documented in http/README.md with an assertion that
demanded the wrong behaviour, so that fixing it would fail loudly. All
four are fixed here and every assertion rewritten to demand the right
behaviour instead.

PATCH /actors/{id} was not a partial update. modify() mapped the payload
onto a brand new Actor, so every field the caller left out was nulled --
wiping first_name/last_name and then failing outright on the NOT NULL
last_update, which is why a partial payload answered 500. It now merges
onto the managed row; ModelMapper already runs with skipNullEnabled, so
absent fields are left alone. This needed @Transactional, which in turn
needed the constructor to stop being private: CGLIB cannot subclass a
class whose only constructor is private, and Spring proxies the bean to
apply the transaction.

POST /customers discarded the payload's store and address and hardcoded
2 and 591, so every customer landed in the same store no matter what was
asked for. Both are @NotNull on the DTO, so @Valid has already rejected a
payload that omits them; they are now resolved through the repositories,
turning an unknown id into a 404 instead of a late foreign key violation.

PUT /customers/{id} answered 201. It replaces an existing customer, so it
answers 200 -- 201 belongs to newCustomer, the only path here that brings
a resource into existence.

GET /films/{id} answered 500 above 32767. film_id is SMALLINT UNSIGNED
(0-65535) but the entity mapped it as Short, while FilmRepository is typed
JpaRepository<Film, Integer>, so ids past Short.MAX_VALUE blew up in the
conversion rather than missing cleanly -- and films 32768-65535 were
unreachable. The entity now uses Integer, matching repository, DTO and
controller. Same class of bug as Staff.staffId, which was Byte against a
TINYINT UNSIGNED column.

None of the four had JUnit coverage, which is how they survived. Added a
test per bug and verified each fails without its fix, then checked all
four against a running instance.

+++ src/main/java/org/sanmarcux/samples/sakila/dao/model/Film.java
@@ -23,7 +23,10 @@ public class Film implements java.io.Serializable {
 
     @Serial
     private static final long serialVersionUID = 1L;
-    private Short filmId;
+    // film_id is SMALLINT UNSIGNED (0-65535); Short tops out at 32767, and FilmRepository
+    // is typed JpaRepository<Film, Integer>, so any id above 32767 blew up converting the
+    // Integer argument into the entity's Short identifier -- a 500 where a 404 belongs.
+    private Integer filmId;
     private Language languageByOriginalLanguageId;
     private Language languageByLanguageId;
     private String title;
@@ -81,11 +84,11 @@ public class Film implements java.io.Serializable {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "film_id", unique = true, nullable = false)
-    public Short getFilmId() {
+    public Integer getFilmId() {
         return this.filmId;
     }
 
-    public void setFilmId(Short filmId) {
+    public void setFilmId(Integer filmId) {
         this.filmId = filmId;
     }
